refactor: move SCORE_FIELD to constants
